requirements.txt
alembic==1.13.1
apispec==5.2.2
Babel==2.9.0
click==8.1.7
Flask==2.0.3
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
flask-smorest==0.38.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.0.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
python-dateutil==2.6.0
pytz==2024.1
six==1.16.0
SQLAlchemy==1.3.0
typing_extensions==4.9.0
webargs==8.4.0
Werkzeug==2.0.0
WTForms==3.1.2
